+++ checker-qual/src/main/java/org/checkerframework/common/returnsreceiver/qual/UnknownThis.java
@@ -21,8 +21,8 @@ import java.lang.annotation.Target;
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE_USE, ElementType.TYPE_PARAMETER})
-@DefaultQualifierInHierarchy
 @SubtypeOf({})
+@DefaultQualifierInHierarchy
 @QualifierForLiterals(LiteralKind.NULL)
 @DefaultFor(value = TypeUseLocation.LOWER_BOUND)
 @InvisibleQualifier