+++ checker/tests/nullness/MisuseProperties.java
@@ -17,11 +17,10 @@ public class MisuseProperties {
         p.put("line.separator", null);
         Hashtable h = p;
         // Error, because HashTable value has NonNull bound.
-        // TODO: false negative. See #365.
-        //// :: error: (argument.type.incompatible) :: warning: [unchecked] unchecked call to
-        //// put(K,V) as a member of the raw type java.util.Hashtable
+        // put(K,V) as a member of the raw type java.util.Hashtable
         // :: warning: [unchecked] unchecked call to put(K,V) as a member of the raw type
         // java.util.Hashtable
+        // :: error: (argument.type.incompatible)
         h.put("line.separator", null);
         // :: error: (argument.type.incompatible)
         System.setProperty("line.separator", null);