+++ framework/src/main/java/org/checkerframework/framework/source/SourceChecker.java
@@ -37,7 +37,6 @@ import org.checkerframework.javacutil.AnnotationProvider;
 import org.checkerframework.javacutil.AnnotationUtils;
 import org.checkerframework.javacutil.BugInCF;
 import org.checkerframework.javacutil.ElementUtils;
-import org.checkerframework.javacutil.SystemUtil;
 import org.checkerframework.javacutil.TreePathUtil;
 import org.checkerframework.javacutil.TreeUtils;
 import org.checkerframework.javacutil.TypeSystemError;
@@ -605,20 +604,6 @@ public abstract class SourceChecker extends AbstractTypeProcessor implements Opt
         // The processingEnvironment field will be set by the superclass's init method.
         // This is used to trigger AggregateChecker's setProcessingEnvironment.
         setProcessingEnvironment(unwrappedEnv);
-
-        // Keep in sync with check in checker-framework/build.gradle .
-        int jreVersion = SystemUtil.jreVersion;
-        if (!hasOption("noJreVersionCheck")
-                && jreVersion != 8
-                && jreVersion != 11
-                && jreVersion != 17
-                && jreVersion != 19) {
-            message(
-                    Kind.NOTE,
-                    "The Checker Framework is tested with JDK 8, 11, 17, and 19."
-                            + " You are using version %d.",
-                    jreVersion);
-        }
 
         if (!hasOption("warnUnneededSuppressionsExceptions")) {
             warnUnneededSuppressionsExceptions = null;