+++ framework/src/test/java/org/checkerframework/framework/testchecker/compound/qual/CCTop.java
@@ -6,7 +6,7 @@ import org.checkerframework.framework.qual.SubtypeOf;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Target;
 
-@SubtypeOf({})
 @Target({ElementType.TYPE_USE, ElementType.TYPE_PARAMETER})
+@SubtypeOf({})
 @DefaultQualifierInHierarchy
 public @interface CCTop {}