+++ framework/src/test/java/org/checkerframework/framework/testchecker/lubglb/quals/LubglbA.java
@@ -12,6 +12,6 @@ import java.lang.annotation.Target;
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE_USE, ElementType.TYPE_PARAMETER})
-@DefaultQualifierInHierarchy
 @SubtypeOf({})
+@DefaultQualifierInHierarchy
 public @interface LubglbA {}